Separate string search algorithms to its own header
